Moving to Webpack 4
